Refactor(P.IModel) from Human-interfaces to Entity-interfaces & float replaced with double
Интерфейсы сделаны более абстрактными, float заменена ради удобства на double

+++ Presentation/IModel/IChangeSpeed.cs
@@ -7,6 +7,6 @@ namespace Presentation.IModel {
         DateTime GetTime();
         void Pause();
         void Play();
-        void ChangeSpeed(float coefficient);
+        void ChangeSpeed(double coefficient);
     }
 }
